parachain staking config in its file
