+++ ownership-chain/node/src/chain_spec/laos.rs
@@ -2,7 +2,7 @@ use super::{get_collator_keys_from_seed, predefined_accounts, Extensions, SAFE_X
 use cumulus_primitives_core::ParaId;
 use fp_evm::GenesisAccount;
 use laos_ownership_runtime::{
-	AccountId, AuraId, Balance, InflationInfo, MinCollatorStake, Precompiles, BLOCKS_PER_YEAR,
+	AccountId, AuraId, Balance, InflationInfo, configs::parachain_staking::MinCollatorStake, Precompiles, BLOCKS_PER_YEAR,
 	REVERT_BYTECODE, UNIT,
 };
 use sc_service::ChainType;
