+++ ownership-chain/runtime/laos/src/configs/mod.rs
@@ -3,6 +3,7 @@ use ownership_parachain_primitives::Balance;
 
 mod multisig;
 mod proxy;
+pub mod parachain_staking;
 
 // Define storage fees as constants for clarity and reuse
 const STORAGE_ITEM_FEE: Balance = 10 * UNIT;
